Move `LoginFieldsMeta` to dedicated type
